requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests==2.31.0
python-dateutil==2.9.0
aiohttp>=3.8.5 